requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pyserial==3.5
pyserial-asyncio==0.6
pandas==2.2.2
pyarrow==17.0.0
pydantic==2.9.2
python-dotenv==1.0.1
PyYAML==6.0.2
matplotlib==3.8.4
numpy==1.26.4
scikit-learn==1.4.2
scipy==1.11.4
websockets==12.0
requests==2.32.3
jupyter==1.0.0
nbformat==5.10.4
nbclient==0.10.0
